+++ addOns/automation/CHANGELOG.md
@@ -3,10 +3,13 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
+## [0.11.0] - 2022-01-19
 ### Added
 - Support for generating stats
 - Form and JSON-based authentication support
+
+### Changed
+- Promoted to beta
 
 ## [0.10.1] - 2022-01-05
 ### Fixed
@@ -115,6 +118,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - First version.
 
+[0.11.0]: https://github.com/zaproxy/zap-extensions/releases/automation-v0.11.0
 [0.10.1]: https://github.com/zaproxy/zap-extensions/releases/automation-v0.10.1
 [0.10.0]: https://github.com/zaproxy/zap-extensions/releases/automation-v0.10.0
 [0.9.0]: https://github.com/zaproxy/zap-extensions/releases/automation-v0.9.0